pyproject.toml
[project]
name = "qfeng"
version = "0.1.0"
description = "Q-FENG: Quantum-Fractal Neurosymbolic Governance — Empirical Validation"
authors = [
    {name = "Ricardo da Silva Kaminski", email = "ricardoskaminski@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "litellm>=1.40",
    "clingo>=5.7",
    "pymupdf>=1.24",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "ipykernel>=6.29",
]
ml = [
    "scikit-learn>=1.5",
    "xgboost>=2.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "torch>=2.3",
]
ui = [
    "streamlit>=1.36",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
